Add `command_palette_hooks` crate
