+++ crates/command_palette_hooks/LICENSE-GPL
@@ -1,0 +1,1 @@
+../../LICENSE-GPL
