+++ crates/vim/Cargo.toml
@@ -17,9 +17,7 @@ anyhow.workspace = true
 async-compat = { version = "0.2.1", "optional" = true }
 async-trait = { workspace = true, "optional" = true }
 collections.workspace = true
-command_palette.workspace = true
-# HACK: We're only depending on `copilot` here for `CommandPaletteFilter`.  See the attached comment on that type.
-copilot.workspace = true
+command_palette_hooks.workspace = true
 editor.workspace = true
 gpui.workspace = true
 language.workspace = true
